blueprints/finetuning_service/src/finetuning-engine/requirements.txt
# Copyright (C) 2025-2026 Intel Corporation
# SPDX-License-Identifier: Apache-2.0

# FastAPI & Web Framework
fastapi==0.115.0
uvicorn[standard]==0.27.0
python-multipart==0.0.22
slowapi==0.1.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0

# Data Validation & Settings
pydantic==2.9.2
pydantic-settings==2.1.0

# HTTP Client
requests>=2.32.2
httpx==0.26.0

# Unsloth & AI Dependencies (Ensure CUDA 12.1+)
unsloth>=2026.1.4
torch>=2.1.0
torchvision>=0.16.0
trl>=0.7.0
transformers>=4.36.0
datasets>=2.16.0

# Compression (for file operations)
pgzip

# Utilities
python-dotenv==1.0.0
psutil==5.9.8

# Keycloak Token Validation
python-jose[cryptography]==3.5.0